Final notes of refactoring
