+++ Assembly-CSharp.csproj
@@ -8,7 +8,8 @@
     <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
     <ProductVersion>10.0.20506</ProductVersion>
     <SchemaVersion>2.0</SchemaVersion>
-    <RootNamespace></RootNamespace>
+    <RootNamespace>
+    </RootNamespace>
     <ProjectGuid>{6777F9E4-8265-838B-7969-121F26B91C33}</ProjectGuid>
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
@@ -56,6 +57,16 @@
     <Analyzer Include="C:\Program Files\Microsoft Visual Studio\2022\Community\Common7\IDE\Extensions\Microsoft\Visual Studio Tools for Unity\Analyzers\Microsoft.Unity.Analyzers.dll" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Assets\GG\Scripts\Unit\Entity.cs" />
+    <Compile Include="Assets\GG\Scripts\Unit\EntityStatsSettings.cs" />
+    <Compile Include="Assets\GG\Scripts\Unit\Firearmr.cs" />
+    <Compile Include="Assets\GG\Scripts\Unit\FirearmStatsSettings.cs" />
+    <Compile Include="Assets\GG\Scripts\Unit\GlobalStatsSettingsRepository.cs" />
+    <Compile Include="Assets\GG\Scripts\Unit\Playerr.cs" />
+    <Compile Include="Assets\GG\Scripts\Unit\Projectile.cs" />
+    <Compile Include="Assets\GG\Scripts\Unit\Stat.cs" />
+    <Compile Include="Assets\GG\Scripts\Unit\StatModifier.cs" />
+    <Compile Include="Assets\GG\Scripts\Unit\UnitStatsSettings.cs" />
     <Compile Include="Assets\Simple Sprite Packer\Scripts\SPSpriteInfo.cs" />
     <Compile Include="Assets\TextMesh Pro\Examples &amp; Extras\Scripts\SkewTextExample.cs" />
     <Compile Include="Assets\GG\Scripts\Unit\Settings\Stats\NumericalStat.cs" />
