--- Assets/GG/Scripts/Shooting/Ammo.cs
@@ -1,6 +1,0 @@
-//using UnityEngine;
-
-//public abstract class Ammo : MonoBehaviour
-//{
-//    public abstract void Launch(Vector2 direction, FirearmSettings firearmSettings);
-//}
