+++ Assets/GG/Scripts/Shooting/Firearm.cs
@@ -1,24 +1,24 @@
-using UnityEngine;
+//using UnityEngine;
 
-public class Firearm : MonoBehaviour
-{
-    //Input: 1
-    private static bool IsFireButtonPressed =>
-        GameObject.FindGameObjectsWithTag("Player")[0].GetComponent<Player>().isFireButtonPressed;
-    //Shooting: 1
-    private Shooter Shooter => GetComponent<Shooter>();
-    //Magazine: 1
-    private Magazine Magazine => GetComponent<Magazine>();
-    //Reload: 1
-    private Reload Reload => GetComponent<Reload>();
-    private void Update()
-    {
-        if (!Reload.IsInProcess && Magazine.IsEmpty)
-        {
-            Magazine.onEmpty?.Invoke();
-            Debug.Log("Magazine of " + gameObject.name + " is empty");
-        }
-        if (!IsFireButtonPressed) return;
-        if (Shooter.CanShoot) Shooter.Shoot();
-    }
-}
+//public class Firearm : MonoBehaviour
+//{
+//    //Input: 1
+//    private static bool IsFireButtonPressed =>
+//        GameObject.FindGameObjectsWithTag("Player")[0].GetComponent<Player>().isFireButtonPressed;
+//    //Shooting: 1
+//    private Firearmr Firearm => GetComponent<Firearmr>();
+//    //Magazine: 1
+//    private Magazine Magazine => GetComponent<Magazine>();
+//    ////Reload: 1
+//    private Reload Reload => GetComponent<Reload>();
+//    private void Update()
+//    {
+//        if (!Firearm.Reload.IsInProcess && Firearm.Magazine.IsEmpty)
+//        {
+//            Magazine.onEmpty?.Invoke();
+//            Debug.Log("Magazine of " + gameObject.name + " is empty");
+//        }
+//        if (!IsFireButtonPressed) return;
+//        if (Firearm.CanShoot) Firearm.Shoot();
+//    }
+//}
