+++ Assets/GG/Scripts/Shooting/ProjectileCreater.cs
@@ -2,13 +2,13 @@ using UnityEngine;
 
 public class ProjectileCreater : MonoBehaviour
 {
-    public GameObject[] CreateProjectiles(FirearmSettings settings)
+    public GameObject[] CreateProjectiles(int singleShotProjectiles, GameObject ammo, Transform firingPoint)
     {
-        Transform firingPoint = settings.transform;
-        GameObject[] projectiles = new GameObject[settings.SingleShotProjectiles];
+        //Transform firingPoint = settings.transform;
+        GameObject[] projectiles = new GameObject[singleShotProjectiles];
         for (int i = 0; i < projectiles.Length; i++)
         {
-            projectiles[i] = Instantiate(settings.Ammo, firingPoint.position, firingPoint.rotation);
+            projectiles[i] = Instantiate(ammo, firingPoint.position, firingPoint.rotation);
         }
         return projectiles;
     }
