+++ Assets/GG/Scripts/Unit/EntityStatsSettings.cs
@@ -1,0 +1,5 @@
+public class EntityStatsSettings
+{
+    public float Size { get; set; }
+    public float MaximumLife { get; set; }
+}
