+++ Assets/GG/Scripts/Unit/FirearmStatsSettings.cs
@@ -1,0 +1,11 @@
+public class FirearmStatsSettings
+{
+    public float Damage { get; set; }
+    public float ShootForce { get; set; }
+    public float ShootsPerSecond { get; set; }
+    public float MaxShootDeflectionAngle { get; set; }
+    public float MagazineSize { get; set; }
+    public float ReloadSpeed { get; set; }
+    public float SingleShootProjectile { get; set; }
+    
+}
