+++ Assets/GG/Scripts/Unit/StatModifier.cs
@@ -1,0 +1,10 @@
+public class StatModifier
+{
+    public StatModifier(OperationType type, float value)
+    {
+        Type = type;
+        Value = value;
+    }
+    public OperationType Type { get; }
+    public float Value { get; }
+}
