+++ Assets/GG/Scripts/Unit/UnitStatsSettings.cs
@@ -1,0 +1,4 @@
+public class UnitStatsSettings : EntityStatsSettings
+{
+    public float Speed { get; set; }
+}
